refactor: rename tsx & css files from MyBuild to PartPicker
